requirements.txt
Appium-Python-Client==4.0.1
pytest==8.3.1
allure-pytest==2.13.5
selene==2.0.0rc9
python-dotenv==1.0.1
requests==2.32.3
selenium==4.26.1
browserstack-local == 1.2.9